// src/core/enum.ts
export enum Gender {
  MALE = 'male',
  FEMALE = 'female',
}

export enum RouterPaths {
  DASHBOARD = 'dashboard',
  EDUCATION = 'education',
  RATING = 'rating',
  SHOP = 'shop',
  SHOP_DETAIL = 'shop-detail',
  ADMIN_PANEL = 'admin-panel',
  FAQ = 'faq',
  RESET_BY_TOKEN = 'reset-password',
  RESET = 'reset',
  USER_ACTIVATION = 'user-activation',
  CODE_CONFIRMATION = 'code-confirmation',
  REGISTRATION = 'registration',
  AUTH_NOTIFICATION = 'auth-notification',
  REGISTRATION_FORM = 'registration-form',
  PROFILE = 'profile',
  TASKS_FEED = 'tasks-feed',
  BATTLES = 'battles',
  BATTLE_INVITATION = 'battles/invitation',
  BATTLE = 'battle',
  PLAY_GAME = 'play-game',
  COMPANY_CHOICE = 'company-choice',
  USE_TERMS = 'terms',
  KNOWLEDGE_BASE = 'knowledge-base',
  CARD_VIEW = 'knowledge-base/card-view',
  NOTIFICATION = 'notification',
  COURSE = 'education/course',
  MY_COURSES = 'education/course/my-courses',
  AVAILABLE_COURSES = 'education/course/available-courses',
  COURSE_MODULE = 'education/course/:courseId/module/:moduleId',
  MODULE_CARD = 'education/course/:courseId/module/:moduleId/card/:cardId',
  MODULE_TEST = 'education/course/:courseId/module/:moduleId/test',
  COURSE_TEST = 'education/course/:courseId/exam',
  REPORTS = 'reports',
}

export enum AdminRouterPaths {
  COMPANY = 'company',
  GROUP_INFO = 'group-info',
  KNOWLEDGE_BASE = 'knowledge-base',
  CARD_CREATION = 'knowledge-base/card-creation',
  CARD_EDIT = 'knowledge-base/card-edit',
  CARD_VIEW = 'knowledge-base/card-view',
  CONTENT = 'content',
  COURSE_INFO = 'content/course',
  COURSE_CREATION_CHOICE = 'content/course/course-creation-choice',
  COURSE_CREATION = 'content/course/course-creation',
  COURSE_EDITION = 'content/course/course-edition',
  TASK_CREATION = 'content/task-creation',
  TASK_EDITION = 'content/task-edition',
  TASK_DETAIL = 'content/task-detail',
  GAME_CREATION = 'content/game/game-creation',
  GAME_EDITION = 'content/game/game-edition',
  COMPANY_SETTINGS = 'company/settings',
  NOTIFICATION_DISTRIBUTION = 'notification-distribution',
  NOTIFICATION_CREATION = 'notification-creation',
  NOTIFICATION_DETAIL = 'notification-detail',
  SHOP = 'shop',
  SHOP_CREATION = 'shop-creation',
  SHOP_EDITION = 'shop-edition',
  MARKETPLACE = 'marketplace',
  PAYMENT = 'payment',
  MARKETPLACE_DETAILED = 'marketplace-detailed',
  MARKETPLACE_EDUCATION_DETAIL = 'marketplace-education',
  REPORTS = 'reports',
  REPORT_PERFORMANCE = 'reports/performance',
  REPORT_ERROR = 'reports/error',
  REPORT_ZEAL = 'reports/zeal',
}

export enum BattleStatus {
  NEW,
  PENDING,
  REVOKED,
  FINISHED,
  EXPIRED,
  CANCELED,
}

export enum BattleUserStatus {
  WIN,
  LOSE,
  PENDING,
}